chore(GlobalLoader): fix review
