+++ packages/react-ui/components/GlobalLoader/usePrevious.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react';
 
 /**
- * Возвраащет предыдущее значение переданной переменной
+ * Возвращет предыдущее значение переданной переменной
  *
  * @param value - значение пропа
  * @returns - возвращает предыдущее значение переданной переменной
@@ -11,6 +11,6 @@ export function usePrevious(value: any) {
   const ref = useRef();
   useEffect(() => {
     ref.current = value;
-  });
+  }, [value]);
   return ref.current;
 }
